Estimation of connectivity indexes from CONEFOR

Use the CONEFOR command line to estimate probabilistic and binary connectivity indexes

EstConefor(
  nodeFile,
  coneforpath = NULL,
  connectionFile,
  typeconnection,
  typepairs,
  index,
  thdist,
  multdist = NULL,
  conprob = NULL,
  onlyoverall = TRUE,
  LA = NULL,
  nrestauration = FALSE,
  prefix = NULL,
  write = NULL
)

Arguments

nodeFile

character. Node file name. If the prefix parameter is used, then use the prefix name.

coneforpath

character. Path to Conefor 2.6 with command line interface (for more details and download http://www.conefor.org/coneforsensinode.html). Example, "C:/Users/coneforWin64.exe".

connectionFile

character. Connection file name. If prefix is true use the prefix name.

typeconnection

character. Indicate if it is distance (dist), probability (prob), adjacency (adj).

typepairs

character. "all" if all pairs of patches are in the distances file and "notall" if only they are the most important connections.

index

character. Select Binary index (CCP, LCP, IIC, BC, BCIIC, NC, NL, H) or Probabilistic (F, AWF, PC, BCPC).

thdist

numeric. Distance threshold to establish connections.

multdist

numeric. Multiple distance threshold to establish connections (meters). The minimum (min) and maximum (max) distance values to be calculated need to be specified, as well as the value by which the distance value must be increased (step) starting from min until max is reached, e.g. multdist = c(min=10000, max=100000, step=10000).

conprob

numeric. If a probabilistic index is selected, indicate the probability of connection under the selected distance threshold, e.g., 0.5 that is 50 percentage of probability connection.

onlyoverall

logical. Estimate only overall index, default TRUE.

LA

numeric. Maximum landscape attribute.

nrestauration

logical. If TRUE Indicates that there are nodes to add to the initial landscape -restauration effect-. The nodeFile needs a third column with values of 0 or 1 indicating for each node whether it already exists in the landscape (1) or is a candidate to be added to the landscape (0).The onlyoverall option have to be FALSE.

prefix

character. Used in case of processing several sites at once.

write

character. Full path direction of the folder.

Value

Four .txt files: "overall_indices", "results_all_EC(PC)", "results_all_overall_indices" y "node_importance".

References

Saura, S. & Torné, J. 2012. Conefor 2.6 user manual (May 2012). Universidad Politécnica de Madrid. Available at www.conefor.org.

Examples

if (FALSE) { # \dontrun{
library(Makurhini)
setwd("~")
EstConefor(nodeFile = "nodes.txt",
          coneforpath = "C:/Users/coneforWin64.exe",
          connectionFile = "dist.txt",
          typeconnection = "dist",
          typepairs = "notall",
          index = "PC",
          thdist = 30000, multdist = NULL,
          conprob = 0.5,
          onlyoverall = FALSE, LA = NULL,
          nrestauration = FALSE,
          prefix= NULL, write = NULL)
} # }