Connectivity indexes in a regular grid

Use the function to compute the Protected Connected (ProtConn), EC, PC or IIC indexes in a regular grid.

MK_Connect_grid(
  nodes,
  area_unit = "ha",
  region = NULL,
  grid = list(hexagonal = TRUE, cellsize = NULL, grid_boundary = FALSE, clip = FALSE,
    tolerance = NULL),
  protconn = TRUE,
  distance_threshold = NULL,
  probability = NULL,
  transboundary = NULL,
  distance = list(type = "centroid"),
  intern = TRUE,
  parallel = NULL
)

Arguments

nodes

sf, SpatVector, SpatialPolygonsDataFrame. Object containing nodes (e.g., habitat patches or fragments) of each time to analyze information. Nodes are spatial data of type vector (class sf, SpatVector, SpatialPolygonsDataFrame). It must be in a projected coordinate system.

area_unit

character. (optional, default = "m2")
. A character indicating the area units when attribute is NULL. Some options are "m2" (the default), "km2", "cm2", or "ha"; See unit_convert for details.

region

object of class sf, SpatialPolygonsDataFrame. Polygon delimiting the region or study area. It must be in a projected coordinate system.

grid

list or object of class sf, SpatialPolygonsDataFrame. Use this parameter to generate a grid indicating its characteristics in a list (see get_grid) or enter the name of an sf class sf or SpatialPolygonsDataFrame with the grid whose coordinate system must be the same as that of the nodes. Example for generating 100 km2 hexagons:
list(hexagonal = TRUE, cellsize = unit_convert(100, "km2", "m2"), grid_boundary = FALSE, clip = FALSE, tolerance = NULL).

protconn

logical. If TRUE then the ProtConn will be estimated; otherwise, the PC index will be estimated.

distance_threshold

A numeric indicating the dispersal distance (meters) of the considered species. If NULL then distance is estimated as the median dispersal distance between nodes. Alternatively, the dispersal_distance function can be used to estimate the dispersal distance using the species home range.

probability

A numeric value indicating the probability that corresponds to the distance specified in the distance_threshold. For example, if the distance_threshold is a median dispersal distance, use a probability of 0.5 (50%). If the distance_threshold is a maximum dispersal distance, set a probability of 0.05 (5%) or 0.01 (1%). Use in case of selecting the "PC" metric. If probability = NULL, then a probability of 0.5 will be used.

transboundary

numeric. Buffer to select transboundary polygones, see MK_ProtConn.

distance

A matrix or list to establish the distance between each pair of nodes. Distance between nodes may be Euclidean distances (straight-line distance) or effective distances (cost distances) by considering the landscape resistance to the species movements.
- If it is a matrix, then the number of columns and rows must be equal to the number of nodes. This distance matrix could be generated by the distancefile function.
- If it is a list of parameters, then it must contain the distance parameters necessary to calculate the distance between nodes. For example, two of the most important parameters: “type” and “resistance”. For "type" choose one of the distances: "centroid" (faster), "edge", "least-cost" or "commute-time". If the type is equal to "least-cost" or "commute-time", then you must use the "resistance" argument. For example: distance(type = "least-cost", resistance = raster_resistance).
To see more arguments see the distancefile function.

intern

logical. Show the progress of the process, default = TRUE. Sometimes the advance process does not reach 100 percent when operations are carried out very quickly.

parallel

numeric. Specify the number of cores to use for parallel processing, default = NULL. Parallelize the function using furrr package.

References

Matt Strimas-Mackey. http://strimas.com/spatial/hexagonal-grids/.
Saura, S., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2017). Protected areas in the world's ecoregions: How well connected are they? Ecological Indicators, 76, 144–158. Saura, S. & Torne, J. (2012). Conefor 2.6. Universidad Politecnica de Madrid. Available at www.conefor.org.
Pascual-Hortal, L. & Saura, S. (2006). Comparison and development of new graph-based landscape connectivity indices: towards the priorization of habitat patches and corridors for conservation. Landscape Ecology, 21(7): 959-967.
Saura, S. & Pascual-Hortal, L. (2007). A new habitat availability index to integrate connectivity in landscape conservation planning: comparison with existing indices and application to a case study. Landscape and Urban Planning, 83(2-3): 91-103.

Examples

if (FALSE) { # \dontrun{
library(Makurhini)
library(sf)
load(system.file("extdata", "Protected_areas.rda",
                package = "Makurhini", mustWork = TRUE))
data("Ecoregions", package = "Makurhini")
ecoregion <- Ecoregions[1,]
plot(st_geometry(ecoregion), col = "#7E6A9F")
#ProtConn
hexagons_priority <- MK_Connect_grid(nodes = Protected_areas,
                                    region = ecoregion,
                                    area_unit = "ha",
                                    grid = list(hexagonal = TRUE,
                                                cellsize = unit_convert(5000, "km2", "m2")),
                                    protconn = TRUE,
                                    distance_threshold = 3000,
                                    probability = 0.5,
                                    transboundary = 6000,
                                    distance = list(type = "centroid"),
                                    intern = TRUE,
                                    parallel = NULL)
hexagons_priority
plot(st_geometry(ecoregion), col = "#7E6A9F")
plot(hexagons_priority["ProtConn"], add = TRUE)
} # }