Multiple Protected Connected (ProtConn)

Estimate Protected Connected (ProtConn) indicator and fractions for multiple regions.

MK_ProtConnMult(
  nodes,
  regions,
  area_unit = "m2",
  distance = list(type = "edge", resistance = NULL),
  distance_thresholds,
  probability,
  transboundary = NULL,
  transboundary_type = "nodes",
  protconn_bound = FALSE,
  geom_simplify = FALSE,
  delta = FALSE,
  CI = "all",
  plot = FALSE,
  write = NULL,
  parallel = NULL,
  intern = TRUE
)

Arguments

nodes

object of class sf, sfc, sfg, SpatVector, spatialPolygonsDataFrame. Spatial data of vector type that normally contains the spatial limits of protected areas. It must be in a projected coordinate system.

regions

object of class sf, sfc, sfg, SpatVector, spatialPolygonsDataFrame. Polygon delimiting the regions or study areas. It must be in a projected coordinate system.

area_unit

character. (optional, default = "m2")
. A character indicating the area units when attribute is NULL. Some options are "m2" (the default), "km2", "cm2", or "ha"; See unit_convert for details.

distance

A list of parameters to calculate distances between node pairs. Supported types are "edge", "least-cost", and "commute-time". For "least-cost" and "commute-time", the resistance argument is required. Examples:

list(type = "edge", keep = 0.5) # uses 50

list(type = "least-cost", resistance = raster_resistance)

See more arguments in distancefile.

distance_thresholds

A numeric indicating the dispersal distance or distances (meters) of the considered species. If NULL then distance is estimated as the median dispersal distance between nodes. Alternatively, the dispersal_distance function can be used to estimate the dispersal distance using the species home range. Can be the same length as the distance_thresholds parameter.

probability

A numeric value indicating the probability that corresponds to the distance specified in the distance_threshold. For example, if the distance_threshold is a median dispersal distance, use a probability of 0.5 (50%). If the distance_threshold is a maximum dispersal distance, set a probability of 0.05 (5%) or 0.01 (1%). Use in case of selecting the "PC" metric. If probability = NULL, then a probability of 0.5 will be used.

transboundary

numeric. Buffer to select polygons (e.g., PAs) in a second round. The selected polygons will have an attribute value = 0, i.e., their contribution for connectivity would be as stepping stones (Saura et al. 2017). One cross-border value or one for each threshold distance can be set.

transboundary_type

character. Two options: "nodes" (methodology from Saura et al. 2017) or "region".
- If it is "nodes", the transboundary is built from the limits of the nodes present in the region (default). - If it is "region", is selected the transboundary is built from the limits of the region.

protconn_bound

logical. If TRUE then the fractions ProtUnConn[design] and ProtConn[bound] will be estimated.

geom_simplify

logical. Slightly simplify the region and nodes geometries.

delta

logical. Estimate the contribution of each node to the ProtConn value in each region.

CI

character. A character vector representing the type of confidence intervals that will be estimated. The value should be any subset of the values c("norm","basic", "stud", "perc", "bca") or "all" which will compute all five types of intervals (see, boot.ci)

plot

logical. Plot the main ProtConn indicators and fractions with their standard deviation, default = FALSE.

write

character. Output folder including the output file name without extension, e.g., "C:/ProtConn/Protfiles".

parallel

numeric. Specify the number of cores to use for parallel processing, default = NULL. Parallelize the function using furrr package and multiprocess plan.

intern

logical. Show the progress of the process, default = TRUE. Sometimes the advance process does not reach 100 percent when operations are carried out very quickly.

Value

For each region:
- Table with the following ProtConn values: ECA, Prot, ProtConn, ProtUnconn, RelConn, ProtUnConn[design], ProtConn[bound], ProtConn[Prot], ProtConn[Within], ProtConn[Contig], ProtConn[Trans], ProtConn[Unprot], ProtConn[Within][land], ProtConn[Contig][land], ProtConn[Unprot][land], ProtConn[Trans][land]

- If plot is not NULL a list is returned with the ProtConn table and a plots. - If delta is TRUE then it returns an sf class object with the importance value (contribution to ProtConn) for each node in the region.

References

Saura, S., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2017). Protected areas in the world’s ecoregions: How well connected are they? Ecological Indicators, 76, 144–158. Saura, S., Bertzky, B., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2018). Protected area connectivity: Shortfalls in global targets and country-level priorities. Biological Conservation, 219(October 2017), 53–67.

Examples

if (FALSE) { # \dontrun{
library(Makurhini)
library(sf)
data("Ecoregions", package = "Makurhini")#'
#For this example, we select the first three columns and the first 10 of the ecoregions
Ecoregions <- Ecoregions[1:15,1:3]
plot(st_geometry(Ecoregions), col = "#7E6A9F")
load(system.file("extdata", "Protected_areas.rda",
                package = "Makurhini", mustWork = TRUE))
#plot(st_geometry(Protected_areas), col="green", add = TRUE) #It may take time to plot all PAs.

test <- MK_ProtConnMult(nodes = Protected_areas,
                        regions = Ecoregions,
                        area_unit = "ha",
                        distance = list(type= "edge", keep = 0.5),
                        distance_thresholds = c(10000, 50000),
                        probability = 0.5, transboundary = 50000,
                        plot = TRUE, write = NULL,
                        parallel = NULL, intern = TRUE)
test
} # }