Estimate the ECA, rECA, dA and dECA.

Equivalent Connected Area (ECA; if the area is used as attribute) or Equivalent Connectivity index (EC)

MK_dECA(
  nodes,
  attribute = NULL,
  area_unit = "m2",
  weighted = FALSE,
  distance = list(type = "centroid", resistance = NULL),
  threshold = NULL,
  metric = "IIC",
  probability = NULL,
  distance_thresholds = NULL,
  LA = NULL,
  plot = FALSE,
  parallel = NULL,
  parallel_mode = 1,
  write = NULL,
  intern = TRUE
)

Arguments

nodes

list. List of objects containing nodes (e.g., habitat patches or fragments) of each time to analyze information. Nodes can be of the following classes:
- Data.frame with at least two columns: the first for node IDs and the second for attributes.
- Spatial data of type vector (class sf, SpatVector, SpatialPolygonsDataFrame). It must be in a projected coordinate system.
- Raster (class RasterLayer, SpatRaster). It must be in a projected coordinate system. The values must be integers representing the ID of each habitat patch or node, with non-habitat areas represented by NA values (see clump or patches).

attribute

character or list. If NULL (only applicable when nodes is of spatial data of vector or raster type) the area of the nodes will be used as the node attribute. The unit of area can be selected using the area_unit parameter. To use an alternative attribute, consider the class type of the object in the nodes parameter:
- If nodes is list of spatial vectors or data.frames, specify the name of the column containing the attribute for the nodes. The column name must be present in each element of the node list.
- If nodes is a list of raster layers then the list must contain numeric vectors with attributes for each period or element of the list of nodes. The first numeric vector in the list must correspond to the first element of the node list. The length of each vector must be equal to the corresponding number of nodes. If the parameter weighted is TRUE then the numeric vector is multiplied by the area of each node to obtain a weighted habitat index.

area_unit

character. (optional, default = "m2")
. A character indicating the area units when attribute is NULL. Some options are "m2" (the default), "km2", "cm2", or "ha"; See unit_convert for details.

weighted

logical. If the nodes are of raster type, you can weight the estimated area of each node by the attribute. When using this parameter the attribute parameter, which must be a vector of length equal to the number of nodes, usually has values between 0 and 1.

distance

A matrix or list to establish the distance between each pair of nodes. Distance between nodes may be Euclidean distances (straight-line distance) or effective distances (cost distances) by considering the landscape resistance to the species movements.
- If it is a matrix, then the number of columns and rows must be equal to the number of nodes. This distance matrix could be generated by the distancefile function.
- If it is a list of parameters, then it must contain the distance parameters necessary to calculate the distance between nodes. For example, two of the most important parameters: “type” and “resistance”. For "type" choose one of the distances: "centroid" (faster), "edge", "least-cost" or "commute-time". If the type is equal to "least-cost" or "commute-time", then you must use the "resistance" argument. For example: distance(type = "least-cost", resistance = raster_resistance). To see more arguments see the distancefile function.
- You can place a list with resistances where there must be one resistance for each time/scenario of patches in the nodes parameter, for example, if nodes has a list of two times then you can use two resistances one for time 1 and one for time 2: distance(type = "least-cost", resistance = list(resistanceT1, resistanceT2)).

threshold

numeric. Pairs of nodes with a distance value greater than this threshold will be discarded in the analysis which can speed up processing.

metric

A character indicating the connectivity metric to use: "PC" (the default and recommended) to calculate the probability of connectivity index, and "IIC" to calculate the binary integral index of connectivity.

probability

A numeric value indicating the probability that corresponds to the distance specified in the distance_threshold. For example, if the distance_threshold is a median dispersal distance, use a probability of 0.5 (50%). If the distance_threshold is a maximum dispersal distance, set a probability of 0.05 (5%) or 0.01 (1%). Use in case of selecting the "PC" metric. If probability = NULL, then a probability of 0.5 will be used.

distance_thresholds

A numeric indicating the dispersal distance or distances (meters) of the considered species. If NULL then distance is estimated as the median dispersal distance between nodes. Alternatively, the dispersal_distance function can be used to estimate the dispersal distance using the species home range. Can be the same length as the distance_thresholds parameter.

LA

numeric. (optional, default = NULL). The maximum landscape attribute, which is the attribute value that would correspond to a hypothetical habitat patch covering all the landscape with the best possible habitat, in which IIC and PC would be equal to 1. For example, if nodes attribute corresponds to the node area, then LA equals total landscape area. If nodes attribute correspond to a quality-weighted area and the quality factor ranges from 0 to 100, LA will be equal to 100 multiplied by total landscape area. The value of LA does not affect at all the importance of the nodes and is only used to calculate the overall landscape connectivity. If no LA value is entered (default) and overall = TRUE or onlyoverall = TRUE, the function will only calculate the numerator of the global connectivity indices and the equivalent connected ECA or EC index.

plot

logical. Also, you can provide the corresponding year for each period of time analyzed, e.g., c("2011", "2014", "2017")

parallel

(optional, default = NULL). A numeric specifying the number of cores to parallelize the index estimation of the PC or IIC index and its deltas.Particularly useful when you have more than 1000 nodes. By default the analyses are not parallelized.

parallel_mode

(optional, default = 1). A numeric indicating the mode of parallelization: Mode 1 (the default option, and recommended for less than 1000 nodes) parallelizes on the connectivity delta estimate, while Mode 2 (recommended for more than 1000 nodes) parallelizes on the shortest paths between vertices estimate.

write

character. Path and name of the output ".csv" file

intern

logical. Show the progress of the process, default = TRUE. Sometimes the advance process does not reach 100 percent when operations are carried out very quickly.

Value

Table with:

- Time: name of the time periods, name of the model or scenario (are taken from the name of the elements of the list of nodes or the plot argument)
- Max. Landscape attribute: maximum landscape attribute
- Habitat area,
- Distance threshold: it is usually a dispersal threshold associated with one or many species and it is set by the user.
- ECA: Equivalent Connected Area or Equivalent Connectivity - Normalized_ECA ( - Normalized_ECA ( - dA: delta Area between times (percentage)
- dECA: delta ECA between times (percentage)
- rECA: relativized ECA (dECA/dA). According to Liang et al. (2021) "an rECA value greater than 1 indicates that habitat changes result in a disproportionately large change in habitat connectivity, while a value lower than 1 indicates connectivity changes due to random habitat changes (Saura et al. 2011; Dilts et al. 2016)".
- dA/dECA comparisons: comparisons between dA and dECA
- Type of change: Type of change using the dECAfun() and the difference between dA and dECA.

References

www.conefor.org

- Saura, S., Estreguil, C., Mouton, C., & Rodríguez-Freire, M. (2011). Network analysis to assess landscape connectivity trends: Application to European forests (1990-2000). Ecological Indicators, 11(2), 407–416. https://doi.org/10.1016/j.ecolind.2010.06.011
Herrera, L. P., Sabatino, M. C., Jaimes, F. R., & Saura, S. (2017). Landscape connectivity and the role of small habitat patches as stepping stones: an assessment of the grassland biome in South America. Biodiversity and Conservation, 26(14), 3465–3479. https://doi.org/10.1007/s10531-017-1416-7
- Liang, J., Ding, Z., Jiang, Z., Yang, X., Xiao, R., Singh, P. B., ... & Hu, H. (2021). Climate change, habitat connectivity, and conservation gaps: a case study of four ungulate species endemic to the Tibetan Plateau. Landscape Ecology, 36(4), 1071-1087.
- Dilts TE, Weisberg PJ, Leitner P, Matocq MD, Inman RD, Nussear KE, Esque TC (2016) Multi-scale connectivity and graph theory highlight critical areas for conservation under climate change. Ecol Appl 26:1223–1237

Examples

if (FALSE) { # \dontrun{
library(Makurhini)
library(sf)

data("list_forest_patches", package = "Makurhini")
data("study_area", package = "Makurhini")
class(list_forest_patches)

Max_attribute <- unit_convert(st_area(study_area), "m2", "ha")

dECA_test <- MK_dECA(nodes= list_forest_patches, attribute = NULL, area_unit = "ha",
                  distance = list(type= "centroid"), metric = "PC",
                  probability = 0.05, distance_thresholds = 5000,
                  LA = Max_attribute, plot= c("1993", "2003", "2007", "2011"),
                  intern = TRUE)
dECA_test

} # }