Cost distances

Cost distances

cost_distances(
  x,
  id,
  type_distance = "least-cost",
  resistance = NULL,
  CostFun = NULL,
  ngh = NULL,
  mask = NULL,
  threshold = NULL,
  geometry_out = NULL,
  bounding_circles = NULL,
  pairwise = TRUE,
  write_table = NULL
)

Arguments

x
object of class sf, sfc, sfg or SpatialPolygons. The shapefile must be in a projected coordinate system.

id
character. Column name with the core id.

type_distance
character. Choose one of the distances: "least-cost" (default) that takes into account obstacles and local friction of the landscape (See, gdistance package); "commute-time" that is analogous to the resistance distance. This distance is based on the random walk theory and calculated using the electrical circuit theory (See, gdistance package).

resistance
raster. Raster object with resistance values (landscape friction).

CostFun
A function to compute the cost to move between cells.The default is the conductance (isotropic cost distance): function(x) 1/mean(x).

ngh
numeric. Neighbour graph (directions) for distance calculations: 4 (von Neu-mann neighbourhood), 8 (Moore neighbourhood) or 16 (king’s and knight’s moves). Default equal 16.

mask
object of class sf, sfc, sfg or SpatialPolygons. For higher processing speed use this option to clip the resistance at the extent of the mask.

threshold
numeric. Distance threshold, pairs of nodes with a distance value above this threshold will be discarded.

geometry_out
numeric. Avoids the "Inf" error and corrects the "Inf" distance values when some spatial geometries are out of the resistance extent. The raster NA values will be replaced by a CONDUCTANCE value that must be provided by the user using this argument, so that it is possible to calculate cost distances for all the pairs of nodes. If NULL, then a Euclidean distance will be calculated to find these distances.

bounding_circles
numeric. If a value is entered, this will create bounding circles around pairs of core areas (recommended for speed, large resistance rasters or pixel resolution < 150 m). Buffer distances are entered in map units. Also, the function is parallelized using and furrr package and multiprocess plan, default = NULL.

pairwise
logical. If TRUE a pairwise table is returned (From, To, distance) otherwise it will be a matrix.

write_table
character. "" indicates output to the console.

Value

cost distance matrix

Details

The function builds on functions out of Jacob van Etten’s ’gdistance’ package.

References

https://cran.r-project.org/web/packages/gdistance/gdistance.pdf