Estimate dispersal distance using home range

Estimate the dispersal distance using the home range and the equation proposed by Bowman, Jaeger and Fahrig (2002)

dispersal_distance(
  home_range = NULL,
  home_range_unit = "ha",
  dispersal_distance_unit = "m"
)

Arguments

home_range

numeric. Species home range

home_range_unit

character. You can set an area unit (e.g., "km2", "cm2", "ha"; see unit_convert). Default equal to hectares "ha".

dispersal_distance_unit

character. You can set an distance unit (e.g., "km", "m", "cm"; see unit_convert). Default equal to meters "m".

Value

Returns a class object data.frame containing the scattering distances for a given home range.

References

Bowman, J., Jaeger, J. A., & Fahrig, L. (2002). Dispersal distance of mammals is proportional to home range size. Ecology, 83(7), 2049-2055.

Examples

dispersal_distance(home_range = 0.6,
                   home_range_unit = "ha",
                   dispersal_distance_unit = "m")
#>                         Type  Distance
#> 1  Median dispersal distance  542.2177
#> 2 maximum dispersal distance 3098.3867