Plot probability of dispersal

Negative exponential dispersal kernel to calculate the probability of dispersal between two nodes. Used in probabilistic indexes (e.g., PC, BCPC, ProtConn etc.)

probability_distance(
  probability,
  median_distance,
  eval_distance = NULL,
  min.prob = 0.01
)

Arguments

probability

numeric. Probability of dispersal at median_distance parameter.

median_distance

numeric. Up to five maximum dispersal distance in km.

eval_distance

numeric. Calculate the probability of dispersal at a specific median distances (km). Available when only one median_distance is used.

min.prob

numeric. Value between 0-1, maximum x axe value.

Examples

if (FALSE) { # \dontrun{
probability_distance(probability= 0.5, median_distance = c(1, 10, 30, 100), min.prob = 0.01)
probability_distance(probability= 0.5, median_distance = 100, eval_distance = 100)
probability_distance(probability= 0, median_distance = 100, eval_distance = 1)
} # }