Convert numeric types from one area unit to another

Convert numeric types from one area unit to another

unit_convert(
  data_unit,
  unit_1 = c("m2", "Dam2", "ha", "km2", "inch2", "foot2", "yard2", "mile2", "m", "km",
    "inch", "foot", "yard", "mile"),
  unit_2 = c("m2", "Dam2", "ha", "km2", "inch2", "foot2", "yard2", "mile2", "m", "km",
    "inch", "foot", "yard", "mile")
)

Arguments

data_unit

numeric. Area value to be converted.

unit_1

character. Select an area unit: "m2" = square meter, "Dam2" = square decametre, "ha" = hectares, "km2" = square kilometer, "inch2" = square inch, "foot2" = square foot, "yard2" = square yard, "mile2" = square mile. Length unit: "m" = meter, "km" = kilometer, "inch", "foot", "yard", "mile"

unit_2

character. Select an area unit: "m2" = square meter, "Dam2" = square decametre, "ha" = hectares, "km2" = square kilometer, "inch2" = square inch, "foot2" = square foot, "yard2" = square yard, "mile2" = square mile. Length unit: "m" = meter, "km" = kilometer, "inch", "foot", "yard", "mile"

Examples

unit_convert(100, "km2", "ha")
#> [1] 10000