Connectivity indexes in a regular grid

Use the function to compute the Protected Connected (ProtConn), EC, PC or IIC indexes in a regular grid.

MK_Connect_grid(
  nodes,
  area_unit = "ha",
  region = NULL,
  grid_param = list(grid_pol = NULL, grid_id = NULL, hexagonal = TRUE, cellsize = NULL,
    grid_boundary = FALSE, clip = FALSE, tolerance = NULL),
  protconn = TRUE,
  distance_threshold = NULL,
  probability = NULL,
  transboundary = NULL,
  distance = list(type = "centroid"),
  intern = TRUE,
  parallel = NULL
)

Arguments

nodes
object of class sf, sfc, sfg or SpatialPolygons. Nodes shapefile, the shapefile must be in a projected coordinate system.

area_unit
character. Attribute area units. You can set an area unit, "Makurhini::unit_covert()" compatible unit ("m2", "Dam2, "km2", "ha", "inch2", "foot2", "yard2", "mile2"). Default equal to hectares "ha".

region
object of class sf, sfc, sfg or SpatialPolygons. Region shapefile, the shapefile must be in a projected coordinate system.

grid_param
list. Parameters of the grid shapefile, see get_grid.Just omit the parameter 'region'. Example, list(grid_pol = NULL, hexagonal = TRUE, grid_id = NULL, cellsize = unit_convert(1000, "km2", "m2"), grid_boundary = FALSE, clip = FALSE, tolerance = NULL).

protconn
logical. If TRUE then the ProtConn will be estimated; otherwise, the PC index will be estimated.

distance_threshold
numeric. Distance threshold to establish connections (crs units, usually meters).

probability
numeric. Probability of direct dispersal between nodes, Default, 0.5, that is 50 percentage of probability connection. If probability = NULL, then it will be the inverse of the mean dispersal distance for the species (1/α; Hanski and Ovaskainen 2000).

transboundary
numeric. Buffer to select transboundary polygones, see MK_ProtConn.

distance
list. See distancefile. Example, list(type= "centroid", resistance = NULL).

intern
logical. Show the progress of the process, default = TRUE. Sometimes the advance process does not reach 100 percent when operations are carried out very quickly.

parallel
numeric. Specify the number of cores to use for parallel processing, default = NULL. Parallelize the function using furrr package and multiprocess plan when there are more than ONE transboundary.

References

Matt Strimas-Mackey. http://strimas.com/spatial/hexagonal-grids/.
Saura, S., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2017). Protected areas in the world's ecoregions: How well connected are they? Ecological Indicators, 76, 144–158. Saura, S. & Torne, J. (2012). Conefor 2.6. Universidad Politecnica de Madrid. Available at www.conefor.org.
Pascual-Hortal, L. & Saura, S. (2006). Comparison and development of new graph-based landscape connectivity indices: towards the priorization of habitat patches and corridors for conservation. Landscape Ecology, 21(7): 959-967.
Saura, S. & Pascual-Hortal, L. (2007). A new habitat availability index to integrate connectivity in landscape conservation planning: comparison with existing indices and application to a case study. Landscape and Urban Planning, 83(2-3): 91-103.

Examples

if (FALSE) {
library(Makurhini)
library(sf)

data("Protected_areas", package = "Makurhini")
data("regions", package = "Makurhini")
ecoregion <- regions[2,]
plot(st_geometry(ecoregion), col="blue")
#ProtConn
hexagons_priority <- MK_Connect_grid(nodes = Protected_areas,
                                    region = ecoregion,
                                    area_unit = "ha",
                                    grid_param = list(hexagonal = TRUE,
                                                      cellsize = unit_convert(1000, "km2", "m2")),
                                    protconn = TRUE,
                                    distance_threshold = 3000,
                                    probability = 0.5,
                                    transboundary = 6000,
                                    distance = list(type = "centroid"),
                                    intern = TRUE,
                                    parallel = NULL)
hexagons_priority
plot(hexagons_priority["ProtConn"])
}