Connectivity indexes in a regular grid

Use the function to compute the Protected Connected (ProtConn)or EC, PC, IIC indexes on a grid.

MK_Connect_grid(grid_pol = NULL, grid_id = NULL,
  grid_type = c("hexagonal", "square"), cellsize = NULL,
  grid_boundary = FALSE, clip = FALSE, tolerance = NULL, nodes,
  region = NULL, attribute = NULL, thintersect = NULL,
  area_unit = "ha", metric = c("ProtConn", "PC"),
  distance = list(type = "centroid"), distance_threshold = NULL,
  probability = NULL, transboundary = NULL, intern = TRUE,
  parallel = FALSE)

Arguments

grid_pol
object of class sf, sfc, sfg or SpatialPolygons. Grid hexagones or squares. The shapefile must be in a projected coordinate system.

grid_id
character. Column name of the grid ID.

grid_type
character. If grid is null you can make a regular grid of "hexagonal" or "square".

cellsize
numeric. Grid area (square kilometers).

grid_boundary
logical.If TRUE, the Incomplete "hexagons" or "squares" in the boundaries of the region will be discarded

clip
logical. If TRUE, the new grid will be clipped to the region area. The operation time will be longer the greater the number of vertices in the polygon of the region, if it is a region with many vertices use the argument "tolerance".

tolerance
numeric. If "clip" is equal to TRUE reduces the number of vertices in the region polygon.

nodes
object of class sf, sfc, sfg or SpatialPolygons. Nodes shapefile, the shapefile must be in a projected coordinate system.

region
object of class sf, sfc, sfg or SpatialPolygons. Region shapefile, the shapefile must be in a projected coordinate system.

attribute
character. Select the nodes attribute: "Intersected area" = Available if the metric argument is equal to ProtConn, it corresponds to intersected Protected areas (default); or another specific column name with the nodes attribute, ideally this attribute mus be an area-weighted index, otherwise the interpretation of the ProtConn or PC metric may change.

thintersect
numeric. Only available if you selected ProtConn as metric. Threshold of intersection in percentage allowed to select or not a target geometry. Example, if thintersect is equal to 90 then a node will be selected only if the intersection between the node and the region is >= 90 percentage. If NULL, thintersect will be 0 (default). (see, "Makurhini::MK_ProtConn")

area_unit
character. Attribute area units. You can set an area unit, "Makurhini::unit_covert()" compatible unit ("m2", "Dam2, "km2", "ha", "inch2", "foot2", "yard2", "mile2"). Default equal to hectares "ha".

metric
character. Choose a connectivity metric: "ProtConn" Protected Connected Land or "PC" Probability of conectivity considering maximum product probabilities.

distance
list. See distancefile(). Example, list(type= "centroid", resistance = NULL).

distance_threshold
numeric. Distance threshold to establish connections (meters).

probability
numeric. Connection probability to the selected distance threshold, e.g., 0.5 (default) that is 50 percentage of probability connection. Use in case of selecting the "PC" metric or "ProtConn".

transboundary
numeric. Buffer to select polygones in a second round, their attribute value = 0, see "Makurhini::MK_ProtConn".

intern
logical. Show the progress of the process, default = TRUE.

parallel
logical. Parallelize the function using furrr package and multiprocess plan, default = FALSE.

References

Matt Strimas-Mackey. http://strimas.com/spatial/hexagonal-grids/.
Saura, S., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2017). Protected areas in the world's ecoregions: How well connected are they? Ecological Indicators, 76, 144–158. Saura, S. & Torne, J. (2012). Conefor 2.6. Universidad Politecnica de Madrid. Available at www.conefor.org.
Pascual-Hortal, L. & Saura, S. (2006). Comparison and development of new graph-based landscape connectivity indices: towards the priorization of habitat patches and corridors for conservation. Landscape Ecology, 21(7): 959-967.
Saura, S. & Pascual-Hortal, L. (2007). A new habitat availability index to integrate connectivity in landscape conservation planning: comparison with existing indices and application to a case study. Landscape and Urban Planning, 83(2-3): 91-103.

Examples

if (FALSE) {
data("Protected_areas", package = "Makurhini")
data("regions", package = "Makurhini")
ecoregion <- regions[2,]
plot(ecoregion, col="blue")

hexagons_priority <- MK_Connect_grid(grid_type = "hexagonal",
                                    cellsize = 10000, grid_boundary = FALSE,
                                    clip = FALSE, nodes = Protected_areas, region = ecoregion,
                                    attribute = "Intersected area", thintersect = NULL,
                                    area_unit = "ha", metric = "ProtConn",
                                    distance = list(type = "centroid"),
                                    distance_threshold = 3000,
                                    probability = 0.5, transboundary = 6000,
                                    intern = TRUE, parallel = FALSE)
hexagons_priority
}