Test the ECA or ProtConn metrics using multiple dispersal distances

Test the ECA or ProtConn metrics using multiple dispersal distances

test_metric_distance(
  nodes,
  attribute = NULL,
  distance1 = NULL,
  distance2 = NULL,
  distance3 = NULL,
  distance4 = NULL,
  metric = "IIC",
  probability = NULL,
  distance_thresholds,
  region = NULL,
  LA = NULL,
  transboundary = NULL,
  area_unit = "ha",
  groups = 3,
  write = NULL,
  intern = TRUE
)

Arguments

nodes
Object of class sf, sfc, sfg or SpatialPolygons.

attribute
character. Column name with the nodes attribute. If NULL, then the nodes area will be estimated and used as the attribute.

distance1
list. Distance parameters. For example: type, resistance,or keep. For "type" choose one of the distances: "centroid" (faster), "edge", "least-cost distance" or "commute distance". If the type is equal to "least-cost distance" or "commute distance", then you have to use the "resistance" argument. "keep" is a numeric value used for higher processing. To See more options consult the help function of distancefile().

distance2
list. see distance1 argument

distance3
list. see distance1 argument

distance4
list. see distance1 argument

metric
character. "IIC" to estimate the ECA using the IIC index, "PC" to estimate the ECA using the PC index or "ProtConn" to estimate the ProtConn indicator using the PC index.

probability
numeric. numeric. Connection probability to the selected distance threshold, e.g., 0.5 (default) that is 50 percentage of probability connection. Use in case of selecting the "PC" metric or "ProtConn". If probability = NULL, then it will be the inverse of the mean dispersal distance for the species (1/α; Hanski and Ovaskainen 2000).

distance_thresholds
numeric. Distances thresholds (minimum 3) to establish connections. For example, one distance: distance_threshold = 30000; two or more specific distances: distance_thresholds = c(30000, 50000, 100000); sequence distances (recommended): distance_thresholds = seq(10000,100000, 10000).

region
object of class sf, sfc, sfg or SpatialPolygons. If metric is equal to "ProtConn" then you must provide a region shapefile.

LA
numeric. Maximum landscape attribute (attribute unit, if attribute is NULL then unit is equal to m2).

transboundary
numeric. Buffer to select polygones in a second round, their attribute value = 0, see Saura et al. 2017. You can set one transboundary value or one per each threshold distance.

area_unit
character. If attribute is NULL you can set an area unit, "Makurhini::unit_covert()" compatible unit(e.g., "m2", "km2", "ha"). Default equal to hectares "ha".

groups
Selected representative threshold distances (distance just before the biggest changes in connectivity metric)

write
character. Folder path and prefix, for example: "C:/Folder/test".

intern
logical. Show the progress of the process, default = TRUE. Sometimes the advance process does not reach 100 percent when operations are carried out very quickly.

References

Correa Ayram, C. A., Mendoza, M. E., Etter, A., & Pérez Salicrup, D. R. (2017). Anthropogenic impact on habitat connectivity: A multidimensional human footprint index evaluated in a highly biodiverse landscape of Mexico. Ecological Indicators, 72, 895–909. https://doi.org/10.1016/j.ecolind.2016.09.007

Examples

if (FALSE) {
library(Makurhini)
library(rgeos)

data("list_forest_patches", package = "Makurhini")
data("study_area", package = "Makurhini")

Max_attribute <- unit_convert(gArea(study_area), "m2", "ha")

test_metric_distance(nodes = list_forest_patches[[1]],
                     distance1 =list(type= "centroid"),
                     distance2 =list(type= "edge"),
                     attribute = NULL, area_unit = "ha",
                     LA = Max_attribute ,
                     distance_thresholds = seq(10000,100000, 10000),
                     groups = 0)

data("Protected_areas", package = "Makurhini")
data("regions", package = "Makurhini")
region <- regions[1,]

test_metric_distance(nodes = Protected_areas,
                     distance1 =list(type= "centroid"),
                     distance2 =list(type= "edge", keep = 0.05),
                     metric = "ProtConn", probability = 0.5,
                     area_unit = "ha",
                     region = region, transboundary = 50000,
                     distance_thresholds = seq(10000,100000, 10000))
}